requirements.txt
antlr4-python3-runtime==4.10
asttokens==2.0.8
backcall==0.2.0
decorator==5.1.1
executing==0.10.0
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prettytable==3.3.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
six==1.16.0
stack-data==0.4.0
traitlets==5.3.0
wcwidth==0.2.5
